pyproject.toml
[tool.poetry]
name = "business-risk-api"
version = "0.1.0"
description = "A Flask-based backend for legal intelligence tool designed to help General Counsels assess business risks"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "business_risk_api"}]

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.0"
firebase-admin = "^6.2.0"
google-cloud-storage = "^2.10.0"
PyMuPDF = "^1.23.0"
python-dotenv = "^1.0.0"
eml-parser = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"] 